# services/go/warehouse-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing warehouses and global stock.",
        "title": "Warehouse Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/health": {
            "get": {
                "description": "Check the uptime status of the service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/stock": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch all inventory across all warehouses (Secure Route)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Get all global stock",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.StockResponse"
                            }
                        }
                    }
                }
            }
        },
        "/stock/{product_id}": {
            "get": {
                "description": "Returns a list of warehouses that have the specified product in stock",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Find warehouses with stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product UUID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/warehouses": {
            "get": {
                "description": "Returns a list of all warehouses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Get all warehouses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Warehouse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a warehouse entry in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Create a new warehouse",
                "parameters": [
                    {
                        "description": "Warehouse Data",
                        "name": "warehouse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Warehouse"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Warehouse"
                        }
                    }
                }
            }
        },
        "/warehouses/{warehouse_id}/stock": {
            "post": {
                "description": "Add or deduct stock for a specific product in a warehouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Update warehouse stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Warehouse UUID",
                        "name": "warehouse_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stock Update Request",
                        "name": "stock",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.StockUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.StockResponse": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "location": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "description": "BIGINT maps to int64",
                    "type": "integer"
                },
                "warehouseId": {
                    "type": "string"
                },
                "warehouseName": {
                    "type": "string"
                }
            }
        },
        "handlers.StockUpdateRequest": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "description": "Can be positive (add) or negative (deduct)",
                    "type": "integer"
                }
            }
        },
        "models.Warehouse": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "location": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "stocks": {
                    "description": "Relationship: One Warehouse has many Stock entries",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.WarehouseStock"
                    }
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
        "models.WarehouseStock": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "product_id": {
                    "description": "Reference to Inventory Service",
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "X-User-Role",
            "in": "header"
        }
    }
}